+++ example/android/app/src/main/kotlin/io/sofluffy/flutter_pty2_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.sofluffy.flutter_pty2_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()